+++ tests/Domain/Account/Reactors/OfferLoanReactorTest.php
@@ -5,10 +5,7 @@ namespace Tests\Domain\Account\Projectors;
 use App\Domain\Account\AccountAggregateRoot;
 use App\Domain\Account\Exceptions\CouldNotSubtractMoney;
 use App\Mail\LoanProposalMail;
-use App\Models\User;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Mail;
-use Illuminate\Support\Str;
 use Tests\TestCase;
 
 class OfferLoanReactorTest extends TestCase
@@ -18,13 +15,7 @@ class OfferLoanReactorTest extends TestCase
     {
         Mail::fake();
 
-        $uuid = Str::uuid();
-
-        $user = User::factory()->create();
-
-        $aggregate = AccountAggregateRoot::retrieve($uuid)
-            ->createAccount('account', $user->id)
-            ->persist();
+        $aggregate = AccountAggregateRoot::retrieve($this->account->uuid);
 
         $aggregate->subtractMoney(5000);
         $aggregate->persist();
@@ -43,6 +34,8 @@ class OfferLoanReactorTest extends TestCase
             $aggregate->subtractMoney(1);
         }, CouldNotSubtractMoney::class);
 
-        Mail::assertSent(LoanProposalMail::class);
+        Mail::assertSent(function (LoanProposalMail $mail) {
+            return $mail->hasTo($this->user->email);
+        });
     }
 }
